requirements.txt
Flask
pandas
transformers
datasets
scikit-learn
torch
python-dotenv